added custom hook
