+++ src/Components/ProfileHeader.jsx
@@ -5,37 +5,28 @@ import { faEdit } from "@fortawesome/free-solid-svg-icons";
 import Button from "./Button";
 import Parse from "parse";
 import TagGenerator from "./TagGenerator";
+import useUserProfile from "../hooks/useUserProfile.js";
 
 const ProfileHeader = () => {
-  // user/setUser state needs to passed down from further up to avoid duplicate code. Is on our to-do list.
-  const [user, setUser] = useState(null);
-  const [loading, setLoading] = useState(true);
 
+  const [user, loading] = useUserProfile();
   const [bannerImg, setBannerImg] = useState(null);
   const [profileImg, setProfileImg] = useState(null);
 
   const bannerRef = useRef(null);
   const profileImgRef = useRef(null);
 
-  // Retrieve "USERS" database object from the logged in "_User" objectId.
+  console.log("rendering ProfileHeader component");
+
   useEffect(() => {
-    const getCurrentUser = async () => {
-      try {
-        const currentUser = Parse.User.current(); // get _User objectId
-        const query = new Parse.Query("USERS");
-        query.equalTo("user", currentUser);
-        const userRecord = await query.first();
-        setUser(userRecord);
-        setBannerImg(userRecord.get("bannerImage").url());
-        setProfileImg(userRecord.get("profileImage").url());
-      } catch (error) {
-        console.log("Error fetching user data: " + error.message);
-      } finally {
-        setLoading(false); // Allows page to be shown.
-      }
-    };
-    getCurrentUser();
-  }, []);
+    console.log("useEffect called for user having changed")
+    console.log(user);
+    console.dir(user);
+    if (user) {
+      setBannerImg(user.get("bannerImage").url());
+      setProfileImg(user.get("profileImage").url());
+    }
+  }, [user]);
 
   const handleBannerEdit = (e) => {
     e.preventDefault();
